fix: re-organize src directory
rm src/workflow and put some files to src/context
