+++ platforms/cloudflare.ts
@@ -1,5 +1,5 @@
-import type { RouteFunction, WorkflowServeOptions } from "../src/workflow";
-import { serve as serveBase } from "../src/workflow";
+import type { RouteFunction, WorkflowServeOptions } from "../src";
+import { serve as serveBase } from "../src";
 
 export type WorkflowBindings = {
   QSTASH_TOKEN: string;
