+++ platforms/svelte.ts
@@ -1,7 +1,7 @@
 import type { RequestHandler } from "@sveltejs/kit";
 
-import type { RouteFunction, WorkflowServeOptions } from "../src/workflow";
-import { serve as serveBase } from "../src/workflow";
+import type { RouteFunction, WorkflowServeOptions } from "../src";
+import { serve as serveBase } from "../src";
 
 /**
  * Serve method to serve a Upstash Workflow in a Nextjs project
